Extract JSON-RPC stdio transport
